Make router pass parameters to callbacks
The router now passes an object to the route callbacks that
contains the values for the path variables that are defined
in the route. The object that is passed to the callbacks
has as keys the variable names, with the passed parameter
as its value. Consider the following route and actual url
that a user requests:

    route: /teams/:id
    request to: /teams/155

This results in the object { id: "155" } being passed
to the callback of the 'teams/:id' route.